Move tests to hyperopttools test file
